Set no_framework_libs: true on apache-harmony-jdwp-tests
apache-harmony-jdwp-tests was previously also built as
apache-harmony-jdwp-tests-hostdex, which is has slightly different
behavior in the build system and defaults to only compiling against
the libcore libraries and not against framework.jar.
apache-harmony-jdwp-tests-hostdex is now an alias for
apache-harmony-jdwp-tests, which breaks builds that don't have
framework.jar in the manifest.  Add no_framework_libs: true,
since the module can clearly compile against only the libcore
libraries.

Test: m checkbuild
Change-Id: Ib0a7959f810fb60cc725ec7e7bcf261b79bb6106