Delete tests that are broken due to missing methods
There are a number of tests that are broken due to missing methods in
some support classes at runtime. The tests build against the version
of the support class in this project which has the methods but they are
run against the version of the class from libcore which does not have
the methods.

This change removes the tests. It has no impact on our coverage as the
tests were not being run. It is needed to allow the duplicate classes
to be removed without breaking the build.

The Support_AvailTest class was removed as it was only used by
SystemProcessTest.

Bug: 131738612
Test: atest CtsLibcoreTestCases
Change-Id: I085db08daabc916d88f1466832269fc614eb7c5f