Avoid conflict with libcore.java.util.logging.OldLoggerTest
If the LoggerTest#testGetLogger_WithParentNormal is run too close in
time to the OldLoggerTest#testGetLogger_WithParent then the second one
will fail. That's because they both use the same name for a logger,
check that logger does not exist before starting and then create the
logger. The only way for the tests to be run successfully is for the
logger that was created to be garbage collected between the runs.

This was detected when investigating another conflict between
logging tests from libcore and external/apache-harmony. The fix is
simply to use a different name for the logger.

Bug: 131738612
Test: atest libcore.java.util.logging \
            org.apache.harmony.logging.tests.java.util.logging

Change-Id: I188e77ecd344e1d33b5c3cf65ad7f15bb87e067c